packages/python/pyproject.toml
[project]
name = "a2a-lite"
version = "0.2.5"
description = "Simplified wrapper for Google's A2A Protocol SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "A2A Lite Contributors" }
]
keywords = ["a2a", "agents", "protocol", "sdk", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "a2a-sdk[http-server]>=0.2.6",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "uvicorn>=0.30.0",
    "httpx>=0.25.0",
    "starlette>=0.40.0",
]

[project.optional-dependencies]
oauth = [
    "pyjwt[crypto]>=2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]

[project.scripts]
a2a-lite = "a2a_lite.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/a2a_lite"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
